ENH: include uncertainty in treatment effects
I am doing a urgent little archelogy project to recover the uncertainty in drug cost and effect that Higashi developed.  I would like confirmation that this is still the best way to handle things for Monday.

The average daily costs of each medication category are:
•    Thiazide-type diuretics = US-bash.009
•    Calcium-channel blockers = US-bash.166
•    Angiotensin-converting enzyme (ACE) inhibitors = US-bash.059
•    Beta blockers = US-bash.048
There are 1,000 draws for the daily costs of medication. These are from PERT distributions.

    TTD – determine effectiveness by taking a draw from N~(8.8,.281)
    CCB – determine effectiveness by taking a draw from N~(8.8, .23)
    ACEI – determine effectiveness by taking a draw from N~(8.5,.281)
    BB – determine effectiveness by taking a draw from N~(9.2, .332)

--Abie