Removed redundant logic check regarding tobacco uptake
We no longer need to avoid applying a non-zero uptake rate to cohorts
where tobacco use is already prevalent. This only occurred in the first
year of the simulation, where the initial prevalence of tobacco use
included the 20-year-old cohorts. These cohorts now consist entirely of
non-smokers, thanks to change in the artifact builder, and so the uptake
rate can be applied as-is.