Create a default native specification for portability
The old native specification was tightly coupled with IHME's cluster,
using custom resources that were not portable to other SGE clusters.
The native specification now defaults to a small subset of flags
that are believed to be standard, and only adds IHME specific flags
if the the cluster name matches expectations.