Change pip freeze usage to use command rather than internal method (#104)
This changes the `pip freeze` usage from an internal method to the command. This allows us to use the `--format` flag that in testing removes the issue in MIC-2580, where local file paths are used instead of versions.

Tested with multiple myeloma run, killed after producing the `requirements.txt`. Tested also with a restart, killed after passing the environment validation step.