gather_results empty data bugfix (#279)
    Category: bugfix
    JIRA issue: MIC-3840

When we filter to an empty population in gather_results, the output of our groupby -> aggregate is an empty dataframe. Vivarium tries to convert this to a Series using squeeze and errors out if it's not a Series. However, an empty dataframe that is squeezed is still a dataframe.
Testing

Added test with filter that would lead to an empty dataframe. Checked that previous code failed test and that updated code passed.