Refactor/transition names (#316)
    Category: refactor
    JIRA issue: MIC-4085

Changes and notes
Define get_transition_names function on BaseDiseaseState
Call this function when getting all transitions in a Disease Model
This is a straight refactor with no logic changes. It's needed to support the change I made in the associated US CVD PR (ihmeuw/vivarium_nih_us_cvd#124) to allow the model to use the standard VPH disease observer

A bunch of whitespcae changes got pulled in by the linter.

Testing
Ran an interactive sim for a few steps and verified that the metrics pipeline was counting transitions as expected. Currently running a parallel sim.